# openapi.json
{
	"openapi": "3.1.0",
	"info": {
		"title": "YOLO & SAM Inference Backend",
		"version": "0.1.0"
	},
	"paths": {
...
		"/rfdetr-models": {
			"get": {
				"tags": ["rfdetr"],
				"summary": "List Rfdetr Models",
				"operationId": "list_rfdetr_models_rfdetr_models_get",
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/RfDetrModelInfo"
									},
									"type": "array",
									"title": "Response List Rfdetr Models Rfdetr Models Get"
								}
							}
						}
					}
				}
			},
			"post": {
				"tags": ["rfdetr"],
				"summary": "Upload Rfdetr Model",
				"operationId": "upload_rfdetr_model_rfdetr_models_post",
				"requestBody": {
					"content": {
						"multipart/form-data": {
							"schema": {
								"$ref": "#/components/schemas/Body_upload_rfdetr_model_rfdetr_models_post"
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RfDetrUploadResponse"
								}
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPValidationError"
								}
							}
						}
					}
				}
			}
		},
		"/rfdetr-models/archive": {
			"post": {
				"tags": ["rfdetr"],
				"summary": "Upload Rfdetr Model Archive",
				"operationId": "upload_rfdetr_model_archive_rfdetr_models_archive_post",
				"requestBody": {
					"content": {
						"multipart/form-data": {
							"schema": {
								"$ref": "#/components/schemas/Body_upload_rfdetr_model_archive_rfdetr_models_archive_post"
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RfDetrUploadResponse"
								}
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPValidationError"
								}
							}
						}
					}
				}
			}
		},
		"/rfdetr-models/{model_name}/weights": {
			"get": {
				"tags": ["rfdetr"],
				"summary": "Download Rfdetr Model Weights",
				"operationId": "download_rfdetr_model_weights_rfdetr_models__model_name__weights_get",
				"parameters": [
					{
						"name": "model_name",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"title": "Model Name"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPValidationError"
								}
							}
						}
					}
				}
			}
		},
		"/rfdetr-models/{model_name}/onnx": {
			"get": {
				"tags": ["rfdetr"],
				"summary": "Export Rfdetr Model Onnx",
				"operationId": "export_rfdetr_model_onnx_rfdetr_models__model_name__onnx_get",
				"parameters": [
					{
						"name": "model_name",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"title": "Model Name"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPValidationError"
								}
							}
						}
					}
				}
			}
		},
		"/rfdetr-models/{model_name}/classes": {
			"get": {
				"tags": ["rfdetr"],
				"summary": "Download Rfdetr Model Classes",
				"operationId": "download_rfdetr_model_classes_rfdetr_models__model_name__classes_get",
				"parameters": [
					{
						"name": "model_name",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"title": "Model Name"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPValidationError"
								}
							}
						}
					}
				}
			}
		},
		"/rfdetr-models/{model_name}": {
			"delete": {
				"tags": ["rfdetr"],
				"summary": "Delete Rfdetr Model",
				"operationId": "delete_rfdetr_model_rfdetr_models__model_name__delete",
				"parameters": [
					{
						"name": "model_name",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"title": "Model Name"
						}
					}
				],
				"responses": {
					"204": {
						"description": "Successful Response"
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPValidationError"
								}
							}
						}
					}
				}
			}
		},
		"/rfdetr-models/{model_name}/rename": {
			"put": {
				"tags": ["rfdetr"],
				"summary": "Rename Rfdetr Model",
				"operationId": "rename_rfdetr_model_rfdetr_models__model_name__rename_put",
				"parameters": [
					{
						"name": "model_name",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"title": "Model Name"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RenameModelRequest"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RfDetrModelInfo"
								}
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPValidationError"
								}
							}
						}
					}
				}
			}
		},
		"/rfdetr-models/{model_name}/annotate": {
			"post": {
				"tags": ["rfdetr"],
				"summary": "Auto Annotate Images Rfdetr",
				"operationId": "auto_annotate_images_rfdetr_rfdetr_models__model_name__annotate_post",
				"parameters": [
					{
						"name": "model_name",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"title": "Model Name"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RfDetrAnnotateRequest"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RfDetrAnnotateResponse"
								}
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPValidationError"
								}
							}
						}
					}
				}
			}
		},
...
		"/training/devices": {
			"get": {
				"tags": ["training"],
				"summary": "List Training Devices",
				"operationId": "list_training_devices_training_devices_get",
				"parameters": [
					{
						"name": "provider",
						"in": "query",
						"required": false,
						"schema": {
							"anyOf": [
								{
									"type": "string"
								},
								{
									"type": "null"
								}
							],
							"title": "Provider"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DevicesResponse"
								}
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPValidationError"
								}
							}
						}
					}
				}
			}
		},
...
		"/training/jobs": {
			"post": {
				"tags": ["training"],
				"summary": "Start Training Job",
				"operationId": "start_training_job_training_jobs_post",
				"requestBody": {
					"required": true,
					"content": {
						"multipart/form-data": {
							"schema": {
								"$ref": "#/components/schemas/Body_start_training_job_training_jobs_post"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TrainingJobDetail"
								}
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPValidationError"
								}
							}
						}
					}
				}
			},
			"get": {
				"tags": ["training"],
				"summary": "List Training Jobs",
				"operationId": "list_training_jobs_training_jobs_get",
				"parameters": [
					{
						"name": "user_id",
						"in": "query",
						"required": false,
						"schema": {
							"anyOf": [
								{
									"type": "string"
								},
								{
									"type": "null"
								}
							],
							"title": "User Id"
						}
					},
					{
						"name": "project_name",
						"in": "query",
						"required": false,
						"schema": {
							"anyOf": [
								{
									"type": "string"
								},
								{
									"type": "null"
								}
							],
							"title": "Project Name"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": {
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/TrainingJobSummary"
									},
									"title": "Response List Training Jobs Training Jobs Get"
								}
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPValidationError"
								}
							}
						}
					}
				}
			}
		},
		"/training/jobs/{job_id}": {
			"get": {
				"tags": ["training"],
				"summary": "Get Training Job",
				"operationId": "get_training_job_training_jobs__job_id__get",
				"parameters": [
					{
						"name": "job_id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"title": "Job Id"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TrainingJobDetail"
								}
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPValidationError"
								}
							}
						}
					}
				}
			}
		},
		"/training/jobs/{job_id}/cancel": {
			"post": {
				"tags": ["training"],
				"summary": "Cancel Training Job",
				"operationId": "cancel_training_job_training_jobs__job_id__cancel_post",
				"parameters": [
					{
						"name": "job_id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"title": "Job Id"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TrainingJobDetail"
								}
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPValidationError"
								}
							}
						}
					}
				}
			}
		},
		"/training/jobs/{job_id}/logs": {
			"get": {
				"tags": ["training"],
				"summary": "Get Training Job Logs",
				"operationId": "get_training_job_logs_training_jobs__job_id__logs_get",
				"parameters": [
					{
						"name": "job_id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"title": "Job Id"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": {
							"text/plain": {
								"schema": {
									"type": "string"
								}
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPValidationError"
								}
							}
						}
					}
				}
			}
		},
		"/training/jobs/{job_id}/artifacts/{filename}": {
			"get": {
				"tags": ["training"],
				"summary": "Get Training Job Artifact",
				"operationId": "get_training_job_artifact_training_jobs__job_id__artifacts__filename__get",
				"parameters": [
					{
						"name": "job_id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"title": "Job Id"
						}
					},
					{
						"name": "filename",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"title": "Filename"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPValidationError"
								}
							}
						}
					}
				}
			}
		},
		"/training/jobs/{job_id}/logs/stream": {
			"get": {
				"tags": ["training"],
				"summary": "Stream Training Job Logs",
				"operationId": "stream_training_job_logs_training_jobs__job_id__logs_stream_get",
				"parameters": [
					{
						"name": "job_id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"title": "Job Id"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPValidationError"
								}
							}
						}
					}
				}
			}
		},
...
	},
	"components": {
		"schemas": {
...
			"Body_start_training_job_training_jobs_post": {
				"properties": {
					"dataset_file": {
						"type": "string",
						"format": "binary",
						"title": "Dataset File"
					},
					"user_id": {
						"type": "string",
						"title": "User Id"
					},
					"project_name": {
						"type": "string",
						"title": "Project Name"
					},
					"output_model_name": {
						"type": "string",
						"title": "Output Model Name"
					},
					"base_model": {
						"type": "string",
						"title": "Base Model"
					},
					"epochs": {
						"type": "integer",
						"title": "Epochs"
					},
					"architecture": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "null"
							}
						],
						"title": "Architecture",
						"default": "yolo"
					},
					"train_percent": {
						"type": "integer",
						"title": "Train Percent"
					},
					"val_percent": {
						"type": "integer",
						"title": "Val Percent"
					},
					"test_percent": {
						"type": "integer",
						"title": "Test Percent"
					},
					"total_images": {
						"type": "integer",
						"title": "Total Images"
					},
					"boxed_images": {
						"type": "integer",
						"title": "Boxed Images"
					},
					"empty_images": {
						"type": "integer",
						"title": "Empty Images"
					},
					"classes_json": {
						"type": "string",
						"title": "Classes Json"
					},
					"imgsz": {
						"anyOf": [
							{
								"type": "integer"
							},
							{
								"type": "null"
							}
						],
						"title": "Imgsz"
					},
					"batch": {
						"anyOf": [
							{
								"type": "integer"
							},
							{
								"type": "null"
							}
						],
						"title": "Batch"
					},
					"workers": {
						"anyOf": [
							{
								"type": "integer"
							},
							{
								"type": "null"
							}
						],
						"title": "Workers"
					},
					"device": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "null"
							}
						],
						"title": "Device",
						"default": "auto"
					},
					"provider": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "null"
							}
						],
						"title": "Provider",
						"default": "local"
					},
					"external_callback_url": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "null"
							}
						],
						"title": "External Callback Url"
					},
					"external_callback_secret": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "null"
							}
						],
						"title": "External Callback Secret"
					},
					"hyperparams_json": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "null"
							}
						],
						"title": "Hyperparams Json"
					}
				},
				"type": "object",
				"required": [
					"dataset_file",
					"user_id",
					"project_name",
					"output_model_name",
					"base_model",
					"epochs",
					"train_percent",
					"val_percent",
					"test_percent",
					"total_images",
					"boxed_images",
					"empty_images",
					"classes_json"
				],
				"title": "Body_start_training_job_training_jobs_post"
			},
			"Body_upload_rfdetr_model_archive_rfdetr_models_archive_post": {
				"properties": {
					"name": {
						"type": "string",
						"title": "Name"
					},
					"variant": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "null"
							}
						],
						"title": "Variant"
					},
					"archive_file": {
						"type": "string",
						"format": "binary",
						"title": "Archive File"
					}
				},
				"type": "object",
				"required": ["name", "archive_file"],
				"title": "Body_upload_rfdetr_model_archive_rfdetr_models_archive_post"
			},
			"Body_upload_rfdetr_model_rfdetr_models_post": {
				"properties": {
					"name": {
						"type": "string",
						"title": "Name"
					},
					"variant": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "null"
							}
						],
						"title": "Variant"
					},
					"weights_file": {
						"type": "string",
						"format": "binary",
						"title": "Weights File"
					},
					"classes_file": {
						"anyOf": [
							{
								"type": "string",
								"format": "binary"
							},
							{
								"type": "null"
							}
						],
						"title": "Classes File"
					}
				},
				"type": "object",
				"required": ["name", "weights_file"],
				"title": "Body_upload_rfdetr_model_rfdetr_models_post"
			},
...
			"DevicesResponse": {
				"properties": {
					"devices": {
						"items": {
							"$ref": "#/components/schemas/DeviceInfo"
						},
						"type": "array",
						"title": "Devices"
					}
				},
				"type": "object",
				"required": ["devices"],
				"title": "DevicesResponse"
			},
...
			"HTTPValidationError": {
				"properties": {
					"detail": {
						"items": {
							"$ref": "#/components/schemas/ValidationError"
						},
						"type": "array",
						"title": "Detail"
					}
				},
				"type": "object",
				"title": "HTTPValidationError"
			},
			"RenameModelRequest": {
				"properties": {
					"new_name": {
						"type": "string",
						"title": "New Name"
					}
				},
				"type": "object",
				"required": ["new_name"],
				"title": "RenameModelRequest"
			},
			"RfDetrAnnotateRequest": {
				"properties": {
					"image_urls": {
						"items": {
							"type": "string"
						},
						"type": "array",
						"title": "Image Urls"
					},
					"conf_threshold": {
						"anyOf": [
							{
								"type": "number"
							},
							{
								"type": "null"
							}
						],
						"title": "Conf Threshold",
						"default": 0.5
					},
					"class_map": {
						"anyOf": [
							{
								"additionalProperties": {
									"type": "string"
								},
								"type": "object"
							},
							{
								"type": "null"
							}
						],
						"title": "Class Map"
					}
				},
				"type": "object",
				"required": ["image_urls"],
				"title": "RfDetrAnnotateRequest"
			},
			"RfDetrAnnotateResponse": {
				"properties": {
					"annotations": {
						"items": {
							"items": {
								"additionalProperties": true,
								"type": "object"
							},
							"type": "array"
						},
						"type": "array",
						"title": "Annotations"
					}
				},
				"type": "object",
				"required": ["annotations"],
				"title": "RfDetrAnnotateResponse"
			},
			"RfDetrModelInfo": {
				"properties": {
					"name": {
						"type": "string",
						"title": "Name"
					},
					"classes": {
						"items": {
							"type": "string"
						},
						"type": "array",
						"title": "Classes"
					},
					"variant": {
						"type": "string",
						"enum": ["nano", "small", "medium", "large"],
						"title": "Variant"
					},
					"resolution": {
						"anyOf": [
							{
								"type": "integer"
							},
							{
								"type": "null"
							}
						],
						"title": "Resolution"
					},
					"date_add": {
						"type": "string",
						"title": "Date Add"
					},
					"size_bytes": {
						"type": "integer",
						"title": "Size Bytes"
					},
					"trained": {
						"type": "boolean",
						"title": "Trained",
						"default": false
					}
				},
				"type": "object",
				"required": ["name", "classes", "variant", "date_add", "size_bytes"],
				"title": "RfDetrModelInfo"
			},
			"RfDetrUploadResponse": {
				"properties": {
					"name": {
						"type": "string",
						"title": "Name"
					},
					"classes": {
						"items": {
							"type": "string"
						},
						"type": "array",
						"title": "Classes"
					},
					"variant": {
						"type": "string",
						"enum": ["nano", "small", "medium", "large"],
						"title": "Variant"
					},
					"message": {
						"type": "string",
						"title": "Message"
					}
				},
				"type": "object",
				"required": ["name", "classes", "variant", "message"],
				"title": "RfDetrUploadResponse"
			},
...
			"TrainingJobDetail": {
				"properties": {
					"id": {
						"type": "string",
						"title": "Id"
					},
					"user_id": {
						"type": "string",
						"title": "User Id"
					},
					"project_name": {
						"type": "string",
						"title": "Project Name"
					},
					"output_model_name": {
						"type": "string",
						"title": "Output Model Name"
					},
					"produced_model_name": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "null"
							}
						],
						"title": "Produced Model Name"
					},
					"base_model": {
						"type": "string",
						"title": "Base Model"
					},
					"architecture": {
						"type": "string",
						"enum": ["yolo", "rfdetr"],
						"title": "Architecture",
						"default": "yolo"
					},
					"status": {
						"type": "string",
						"enum": ["queued", "running", "cancelling", "cancelled", "completed", "failed", "interrupted"],
						"title": "Status"
					},
					"provider": {
						"type": "string",
						"enum": ["local", "runpod"],
						"title": "Provider",
						"default": "local"
					},
					"epochs": {
						"type": "integer",
						"title": "Epochs"
					},
					"imgsz": {
						"anyOf": [
							{
								"type": "integer"
							},
							{
								"type": "null"
							}
						],
						"title": "Imgsz"
					},
					"batch": {
						"anyOf": [
							{
								"type": "integer"
							},
							{
								"type": "null"
							}
						],
						"title": "Batch"
					},
					"workers": {
						"anyOf": [
							{
								"type": "integer"
							},
							{
								"type": "null"
							}
						],
						"title": "Workers"
					},
					"device": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "null"
							}
						],
						"title": "Device"
					},
					"device_name": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "null"
							}
						],
						"title": "Device Name"
					},
					"total_images": {
						"type": "integer",
						"title": "Total Images"
					},
					"boxed_images": {
						"type": "integer",
						"title": "Boxed Images"
					},
					"empty_images": {
						"type": "integer",
						"title": "Empty Images"
					},
					"classes": {
						"items": {
							"type": "string"
						},
						"type": "array",
						"title": "Classes"
					},
					"split": {
						"$ref": "#/components/schemas/SplitConfig"
					},
					"created_at": {
						"type": "string",
						"title": "Created At"
					},
					"started_at": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "null"
							}
						],
						"title": "Started At"
					},
					"completed_at": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "null"
							}
						],
						"title": "Completed At"
					},
					"metrics": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/TrainingMetricsSummary"
							},
							{
								"type": "null"
							}
						]
					},
					"error": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "null"
							}
						],
						"title": "Error"
					},
					"artifacts": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/TrainingArtifacts"
							},
							{
								"type": "null"
							}
						]
					},
					"current_epoch": {
						"anyOf": [
							{
								"type": "integer"
							},
							{
								"type": "null"
							}
						],
						"title": "Current Epoch"
					},
					"total_epochs": {
						"anyOf": [
							{
								"type": "integer"
							},
							{
								"type": "null"
							}
						],
						"title": "Total Epochs"
					},
					"runpod_endpoint_id": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "null"
							}
						],
						"title": "Runpod Endpoint Id"
					},
					"runpod_job_id": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "null"
							}
						],
						"title": "Runpod Job Id"
					},
					"logs_tail": {
						"items": {
							"type": "string"
						},
						"type": "array",
						"title": "Logs Tail"
					}
				},
				"type": "object",
				"required": [
					"id",
					"user_id",
					"project_name",
					"output_model_name",
					"base_model",
					"status",
					"epochs",
					"total_images",
					"boxed_images",
					"empty_images",
					"split",
					"created_at"
				],
				"title": "TrainingJobDetail"
			},
...
			"ValidationError": {
				"properties": {
					"loc": {
						"items": {
							"anyOf": [
								{
									"type": "string"
								},
								{
									"type": "integer"
								}
							]
						},
						"type": "array",
						"title": "Location"
					},
					"msg": {
						"type": "string",
						"title": "Message"
					},
					"type": {
						"type": "string",
						"title": "Error Type"
					}
				},
				"type": "object",
				"required": ["loc", "msg", "type"],
				"title": "ValidationError"
			},
...
		}
	}
}